Add MapStruct dependency; Refactor unit tests;
Closes #20
