+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/CategoryTest.java
@@ -1,26 +1,18 @@
-package ru.aahzbrut.reciperestapi.entities.domain;
+package ru.aahzbrut.reciperestapi.domain.entities;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
-import ru.aahzbrut.reciperestapi.domain.entities.Category;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 @Slf4j
-class CategoryTest {
+class CategoryTest extends BaseTestEntity {
 
     private static final Long firstId = 1L;
     private static final Long secondId = 2L;
     private static final String firstName = "First recipe";
     private static final String secondName = "Second recipe";
-
-    @Test
-    void npeDefence() {
-        TestUtils.npeDefence(Category.class);
-        int one = 1;
-        assertEquals(1, one);
-    }
 
     @Test
     void equalsTest() {
