+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/IngredientTest.java
@@ -1,0 +1,5 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class IngredientTest extends BaseTestEntity {
+
+}
