+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/NoteTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class NoteTest extends BaseTestEntity {
+}
