+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/RecipeStepTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class RecipeStepTest extends BaseTestEntity {
+}
