+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/RecipeTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class RecipeTest extends BaseTestEntity {
+}
