+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/StepIngredientTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class StepIngredientTest extends BaseTestEntity {
+}
