+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/StepTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class StepTest extends BaseTestEntity {
+}
