+++ src/test/java/ru/aahzbrut/reciperestapi/domain/entities/UOMTest.java
@@ -1,0 +1,4 @@
+package ru.aahzbrut.reciperestapi.domain.entities;
+
+class UOMTest extends BaseTestEntity {
+}
